Correct memory addresses and autosplitter starts, pauses and unpauses

+++ readmem.hpp
@@ -22,13 +22,13 @@ using std::cin;
 class ReadMemory
 {
     public:
-        uint32_t readMem(uint32_t mem, int pid, int memAddress, struct iovec memLocal, struct iovec memRemote)
+        uint32_t readMem(uint32_t mem, int pid, uint64_t memAddress, struct iovec memLocal, struct iovec memRemote)
         {
             uint32_t value;  // Variable to store the read value
 
             memLocal.iov_base = &value;  // Use the value variable
             memLocal.iov_len = sizeof(value);
-            memRemote.iov_base = (void*)(uintptr_t)memAddress;
+            memRemote.iov_base = (void*)memAddress;
             memRemote.iov_len = sizeof(value);
 
             ssize_t memNread = process_vm_readv(pid, &memLocal, 1, &memRemote, 1, 0);
